+++ src/test/java/RegistrationTest.java
@@ -4,44 +4,49 @@ import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.stellarburgers.Login;
-import org.stellarburgers.Registration;
-import org.stellarburgers.UserGenerator;
+import org.stellarburgers.sitepages.Login;
+import org.stellarburgers.sitepages.Registration;
+import org.stellarburgers.user.User;
+import org.stellarburgers.user.UserGenerator;
 
 public class RegistrationTest {
+
     private WebDriver driver;
+    private User user;
+    private Login client;
+    private Registration registration;
+    private String accessToken;
 
     @Before
     public void setUp() {
-        System.setProperty("web-driver.chrome.driver", "C:/WebDriver/bin/chromedriver.exe");
+        System.setProperty("web-driver.chrome.driver", "src/properties/chromedriver.exe");
         driver = new ChromeDriver();
+        user = new UserGenerator().random();
+        client = new Login(driver);
+        registration = new Registration(driver);
     }
 
     @After
     public void teardown() {
         driver.quit();
+        if (accessToken != null) {
+            client.delete(accessToken);
+        }
     }
-
 
     @Test
     @DisplayName("Verification of successful registration")
     public void userRegistration() {
-        var user = new UserGenerator().random();
-        Login client = new Login(driver);
-        Registration registration = new Registration(driver);
         registration
                 .open()
                 .register(user);
-        String accessToken = client.Login(user).extract().path("accessToken");
-        client.delete(accessToken);
+        accessToken = client.Login(user).extract().path("accessToken");
     }
 
 
     @Test
     @DisplayName("Checking the minimum number of characters in the password — six characters")
     public void errorEnteringIncorrectPassword() {
-        var user = new UserGenerator().random();
-        Registration registration = new Registration(driver);
         user.setPassword("12345");
         registration
                       .open()
